wifi: Create new WifiConfiguration for savedConfig in addOrUpdateNetworkNative
With commit 5871e49caae132f154064807a864c7601c754da7, configurations
shall be set only if modified (in addOrUpdateNetworkNative).
To check if modified, savedConfig referenced the framework cache config
(from mConfiguredNetworks), then overwrite it with existing config from
supplicant.
If the config (to be updated) referenced the same framework cache config,
configurations has been overwritten hence will not be set.

In this specific instance, framework tries to DISABLE_TLS_1_2, but this
configuration shall never be set as the corresponding configuration is
always overwritten by the one in wpa_supplicant (which does not have
this configuration set).

This change is to create new WifiConfigration for savedConfig.

Change-Id: Id2ef886d4cbba840140a39f3cb4f00f992cf9f47
CRs-Fixed: 1033747